ml-service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.3
numpy==1.26.4
scikit-learn==1.5.0
joblib==1.4.2
python-dotenv==1.0.1
httpx==0.27.0
langchain==0.2.6
langchain-openai==0.1.13
langchain-core==0.2.10
